add .editorconfig and move source to com.meilisearch.sdk namespace add gradle setup for junit5
