+++ .github/workflows/gradle.yml
@@ -4,23 +4,23 @@
 name: Java CI with Gradle
 
 on:
-  push:
-    branches: [ master, develop ]
-  pull_request:
-    branches: [ master ]
+    push:
+        branches: [ master, develop ]
+    pull_request:
+        branches: [ master ]
 
 jobs:
-  build:
+    build:
 
-    runs-on: ubuntu-latest
+        runs-on: ubuntu-latest
 
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up JDK 1.8
-      uses: actions/setup-java@v1
-      with:
-        java-version: 1.8
-    - name: Grant execute permission for gradlew
-      run: chmod +x gradlew
-    - name: Build with Gradle
-      run: ./gradlew build
+        steps:
+            -   uses: actions/checkout@v2
+            -   name: Set up JDK 1.8
+                uses: actions/setup-java@v1
+                with:
+                    java-version: 1.8
+            -   name: Grant execute permission for gradlew
+                run: chmod +x gradlew
+            -   name: Build with Gradle
+                run: ./gradlew build
