+++ src/main/java/com/meilisearch/sdk/Config.java
@@ -1,4 +1,4 @@
-package meilisearch;
+package com.meilisearch.sdk;
 
 /**
  * Meilisearch configuration
@@ -9,6 +9,7 @@ public class Config {
 
 	/**
 	 * Create a configuration without an API key
+	 *
 	 * @param hostUrl URL of the Meilisearch instance
 	 */
 	public Config(String hostUrl) {
@@ -17,8 +18,9 @@ public class Config {
 
 	/**
 	 * Create a configuration with an API key
+	 *
 	 * @param hostUrl URL of the Meilisearch instance
-	 * @param apiKey API key to pass to the header of requests sent to Meilisearch
+	 * @param apiKey  API key to pass to the header of requests sent to Meilisearch
 	 */
 	public Config(String hostUrl, String apiKey) {
 		this.hostUrl = hostUrl;
