+++ src/main/java/com/meilisearch/sdk/SearchRequest.java
@@ -1,4 +1,4 @@
-package meilisearch;
+package com.meilisearch.sdk;
 
 /**
  * Search request query string builder
@@ -14,32 +14,32 @@ class SearchRequest {
 	String filters;
 	boolean matches;
 
-	SearchRequest (String q) {
+	SearchRequest(String q) {
 		this(q, 0);
 	}
 
-	SearchRequest (String q, int offset) {
+	SearchRequest(String q, int offset) {
 		this(q, offset, 20);
 	}
 
-	SearchRequest (String q, int offset, int limit) {
+	SearchRequest(String q, int offset, int limit) {
 		this(q, offset, limit, "*");
 	}
 
-	SearchRequest (String q, int offset, int limit, String attributesToRetrieve) {
+	SearchRequest(String q, int offset, int limit, String attributesToRetrieve) {
 		this(q, offset, limit, attributesToRetrieve, null, 200, null, null, false);
 	}
 
-	SearchRequest (String q,
-	               int offset,
-	               int limit,
-	               String attributesToRetrieve,
-	               String attributesToCrop,
-	               int cropLength,
-	               String attributesToHighlight,
-	               String filters,
-	               boolean matches) {
-		this.q = q.replaceAll("\\s+?","%20");
+	SearchRequest(String q,
+				  int offset,
+				  int limit,
+				  String attributesToRetrieve,
+				  String attributesToCrop,
+				  int cropLength,
+				  String attributesToHighlight,
+				  String filters,
+				  boolean matches) {
+		this.q = q.replaceAll("\\s+?", "%20");
 		this.offset = offset;
 		this.limit = limit;
 		this.attributesToRetrieve = attributesToRetrieve;
