+++ src/test/java/com/meilisearch/sdk/IndexesTest.java
@@ -1,11 +1,10 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package meilisearch;
+package com.meilisearch.sdk;
 
 import org.junit.Before;
-import org.junit.Test;
-
+import org.junit.jupiter.api.Test;
 
 public class IndexesTest {
 
@@ -23,7 +22,7 @@ public class IndexesTest {
 
 	@Test
 	public void getIndexes() throws Exception {
-		MeiliSearchIndex[] meilisearchIndices = ms.getIndexList();
+		Index[] meilisearchIndices = ms.getIndexList();
 		for (int a = 0; a < meilisearchIndices.length; a++) {
 			System.out.println(meilisearchIndices[a]);
 		}
@@ -33,7 +32,7 @@ public class IndexesTest {
 	@Test
 	public void getIndex() throws Exception {
 		// TODO: input uid for test
-		MeiliSearchIndex meilisearchIndex = ms.getIndex("movies");
+		Index meilisearchIndex = ms.getIndex("movies");
 		System.out.println(meilisearchIndex);
 	}
